+++ sizing.v
@@ -1,7 +1,7 @@
 module gui
 
 // SizingType describes the three sizing modes of GUI
-pub enum SizingType {
+pub enum SizingType as u8 {
 	fit   // element fits to content
 	fill  // element fills to parent (grows or shrinks)
 	fixed // element unchanged
